rename osm trail, and use it.
