+++ routes/api/trails.js
@@ -2,11 +2,11 @@
 
 var express = require('express')
 var router = module.exports = express.Router()
-var Trail = require('../../models/trail')
+var Trail = require('../../models/osm_trail')
 
 
 router.route('/') // this is /api/trails
-    .get(function (req, res, next) {
+    .get(function (req, res) {
       // query string will need updating.
       var query = {}
       if (req.query.center) {
@@ -15,7 +15,7 @@ router.route('/') // this is /api/trails
           geometry: {
             $near: {
               $geometry: {
-                type: "Point",
+                type: 'Point',
                 coordinates: pointArray
               }
             }
@@ -35,7 +35,7 @@ router.route('/') // this is /api/trails
        )
     })
 router.route('/:id') // this is /api/trails/:id
-    .get(function(req, res, next) {
+    .get(function(req, res) {
       Trail.findOne({_id: req.params.id})
         .exec()
         .then(
